feat: 일부 DTO builder 추가
