--- src/main/java/com/example/ootd/domain/clothes/ClothesType.java
@@ -1,6 +1,0 @@
-package com.example.ootd.domain.clothes;
-
-public enum ClothesType {
-  TOP, BOTTOM, DRESS, OUTER, UNDERWEAR,
-  ACCESSORY, SHOES, SOCKS, HAT, BAG, SCARF, ETC
-}
