+++ src/main/java/com/example/ootd/domain/clothes/dto/data/ClothesAttributeDto.java
@@ -1,0 +1,12 @@
+package com.example.ootd.domain.clothes.dto.data;
+
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record ClothesAttributeDto(
+    UUID definitionId,
+    String value
+) {
+
+}
