+++ src/main/java/com/example/ootd/domain/clothes/dto/data/OotdDto.java
@@ -1,0 +1,17 @@
+package com.example.ootd.domain.clothes.dto.data;
+
+import com.example.ootd.domain.clothes.entity.ClothesType;
+import java.util.List;
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record OotdDto(
+    UUID clothesId,
+    String name,
+    String imageUrl,
+    ClothesType type,
+    List<ClothesAttributeWithDefDto> attributes
+) {
+
+}
