+++ src/main/java/com/example/ootd/domain/clothes/dto/data/RecommendationDto.java
@@ -1,0 +1,14 @@
+package com.example.ootd.domain.clothes.dto.data;
+
+import java.util.List;
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record RecommendationDto(
+    UUID weatherId,
+    UUID userId,
+    List<ClothesDto> clothes
+) {
+
+}
