+++ src/main/java/com/example/ootd/domain/clothes/dto/request/ClothesAttributeDefCreateRequest.java
@@ -1,0 +1,10 @@
+package com.example.ootd.domain.clothes.dto.request;
+
+import java.util.List;
+
+public record ClothesAttributeDefCreateRequest(
+    String name,
+    List<String> selectableValue
+) {
+
+}
