+++ src/main/java/com/example/ootd/domain/clothes/dto/request/ClothesAttributeDefUpdateRequest.java
@@ -1,0 +1,10 @@
+package com.example.ootd.domain.clothes.dto.request;
+
+import java.util.List;
+
+public record ClothesAttributeDefUpdateRequest(
+    String name,
+    List<String> selectableValue
+) {
+
+}
