+++ src/main/java/com/example/ootd/domain/clothes/dto/request/ClothesUpdateRequest.java
@@ -1,0 +1,13 @@
+package com.example.ootd.domain.clothes.dto.request;
+
+import com.example.ootd.domain.clothes.dto.data.ClothesAttributeDto;
+import com.example.ootd.domain.clothes.entity.ClothesType;
+import java.util.List;
+
+public record ClothesUpdateRequest(
+    String name,
+    ClothesType type,
+    List<ClothesAttributeDto> attributes
+) {
+
+}
