+++ src/main/java/com/example/ootd/domain/clothes/entity/ClothesType.java
@@ -1,0 +1,6 @@
+package com.example.ootd.domain.clothes.entity;
+
+public enum ClothesType {
+  TOP, BOTTOM, DRESS, OUTER, UNDERWEAR,
+  ACCESSORY, SHOES, SOCKS, HAT, BAG, SCARF, ETC
+}
