+++ src/main/java/com/example/ootd/domain/clothes/service/ClothesService.java
@@ -1,0 +1,6 @@
+package com.example.ootd.domain.clothes.service;
+
+public interface ClothesService {
+
+
+}
