--- src/main/java/com/example/ootd/domain/feed/FeedLike.java
@@ -1,61 +1,0 @@
-package com.example.ootd.domain.feed;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
-import java.time.LocalDateTime;
-import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-/**
- * 피드 좋아요 엔티티
- */
-@Entity
-@Table(name = "feed_likes",
-    uniqueConstraints = {
-        @UniqueConstraint(
-            name = "pk_feed_user",
-            columnNames = {"feed_id", "user_id"}
-        )
-    })
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@EntityListeners(AuditingEntityListener.class)
-public class FeedLike {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.UUID)
-  @Column(nullable = false, updatable = false)
-  private UUID id;  // pk
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "feed_id", nullable = false)
-  private Feed feed;  // 좋아요 누른 피드
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "user_id", nullable = false)
-  private User user;  // 좋아요 누른 사용자
-
-  @Column(nullable = false, updatable = false)
-  @CreatedDate
-  private LocalDateTime createdAt;  // 생성일
-
-  @Builder
-  public FeedLike(Feed feed, User user) {
-    this.feed = feed;
-    this.user = user;
-  }
-}
