+++ src/main/java/com/example/ootd/domain/feed/dto/data/AuthorDto.java
@@ -1,0 +1,13 @@
+package com.example.ootd.domain.feed.dto.data;
+
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record AuthorDto(
+    UUID userId,
+    String name,
+    String profileImageUrl
+) {
+
+}
