+++ src/main/java/com/example/ootd/domain/feed/dto/data/FeedDto.java
@@ -1,0 +1,23 @@
+package com.example.ootd.domain.feed.dto.data;
+
+import com.example.ootd.domain.clothes.dto.data.OotdDto;
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.UUID;
+import lombok.Builder;
+
+@Builder
+public record FeedDto(
+    UUID id,
+    LocalDateTime createdAt,
+    LocalDateTime updatedAt,
+    User author,
+    Weather weather,
+    List<OotdDto> ootds,
+    String content,
+    long likeCount,
+    int commentCount,
+    boolean likedByMe
+) {
+
+}
