+++ src/main/java/com/example/ootd/domain/feed/dto/request/CommentCreateRequest.java
@@ -1,0 +1,11 @@
+package com.example.ootd.domain.feed.dto.request;
+
+import java.util.UUID;
+
+public record CommentCreateRequest(
+    UUID feadId,
+    UUID authorId,
+    String content
+) {
+
+}
