+++ src/main/java/com/example/ootd/domain/feed/dto/request/FeedCreateRequest.java
@@ -1,0 +1,13 @@
+package com.example.ootd.domain.feed.dto.request;
+
+import java.util.List;
+import java.util.UUID;
+
+public record FeedCreateRequest(
+    UUID authorId,
+    UUID weatherId,
+    List<UUID> clothesIds,
+    String content
+) {
+
+}
