+++ src/main/java/com/example/ootd/domain/feed/dto/request/FeedUpdateRequest.java
@@ -1,0 +1,7 @@
+package com.example.ootd.domain.feed.dto.request;
+
+public record FeedUpdateRequest(
+    String content
+) {
+
+}
