+++ src/main/java/com/example/ootd/domain/feed/service/FeedService.java
@@ -1,0 +1,5 @@
+package com.example.ootd.domain.feed.service;
+
+public interface FeedService {
+
+}
