--- src/main/java/com/example/ootd/domain/image/Image.java
@@ -1,45 +1,0 @@
-package com.example.ootd.domain.image;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-/**
- * 이미지 엔티티
- */
-@Entity
-@Table(name = "images")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@EntityListeners(AuditingEntityListener.class)
-public class Image {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.UUID)
-  @Column(nullable = false, updatable = false)
-  private UUID id;  // pk
-
-  @Column(nullable = false, unique = true)
-  private String url; // 이미지 url
-
-  @Column(nullable = false, updatable = false)
-  @CreatedDate
-  private LocalDateTime createdAt;  // 생성일
-
-  @Builder
-  public Image(String url) {
-    this.url = url;
-  }
-}
