--- src/main/java/com/example/ootd/dto/clothes/data/ClothesAttributeDto.java
@@ -1,10 +1,0 @@
-package com.example.ootd.dto.clothes.data;
-
-import java.util.UUID;
-
-public record ClothesAttributeDto(
-    UUID definitionId,
-    String value
-) {
-
-}
