--- src/main/java/com/example/ootd/dto/clothes/data/OotdDto.java
@@ -1,15 +1,0 @@
-package com.example.ootd.dto.clothes.data;
-
-import com.example.ootd.domain.clothes.ClothesType;
-import java.util.List;
-import java.util.UUID;
-
-public record OotdDto(
-    UUID clothesId,
-    String name,
-    String imageUrl,
-    ClothesType type,
-    List<ClothesAttributeWithDefDto> attributes
-) {
-
-}
