--- src/main/java/com/example/ootd/dto/clothes/request/ClothesAttributeDefCreateRequest.java
@@ -1,10 +1,0 @@
-package com.example.ootd.dto.clothes.request;
-
-import java.util.List;
-
-public record ClothesAttributeDefCreateRequest(
-    String name,
-    List<String> selectableValue
-) {
-
-}
