--- src/main/java/com/example/ootd/dto/clothes/request/ClothesAttributeDefUpdateRequest.java
@@ -1,10 +1,0 @@
-package com.example.ootd.dto.clothes.request;
-
-import java.util.List;
-
-public record ClothesAttributeDefUpdateRequest(
-    String name,
-    List<String> selectableValue
-) {
-
-}
