--- src/main/java/com/example/ootd/dto/feed/data/AuthorDto.java
@@ -1,11 +1,0 @@
-package com.example.ootd.dto.feed.data;
-
-import java.util.UUID;
-
-public record AuthorDto(
-    UUID userId,
-    String name,
-    String profileImageUrl
-) {
-
-}
