--- src/main/java/com/example/ootd/dto/feed/request/CommentCreateRequest.java
@@ -1,11 +1,0 @@
-package com.example.ootd.dto.feed.request;
-
-import java.util.UUID;
-
-public record CommentCreateRequest(
-    UUID feadId,
-    UUID authorId,
-    String content
-) {
-
-}
