--- src/main/java/com/example/ootd/dto/feed/request/FeedCreateRequest.java
@@ -1,13 +1,0 @@
-package com.example.ootd.dto.feed.request;
-
-import java.util.List;
-import java.util.UUID;
-
-public record FeedCreateRequest(
-    UUID authorId,
-    UUID weatherId,
-    List<UUID> clothesIds,
-    String content
-) {
-
-}
