--- src/main/java/com/example/ootd/dto/feed/request/FeedUpdateRequest.java
@@ -1,7 +1,0 @@
-package com.example.ootd.dto.feed.request;
-
-public record FeedUpdateRequest(
-    String content
-) {
-
-}
